task-1-kubernetes/requirements.txt
azure-eventhub==5.11.7
